+++ tests/python_tests/common.py
@@ -14,6 +14,8 @@ from transformers import GenerationConfig as HFGenerationConfig
 from typing import List, Tuple, Callable
 
 from utils.generation_config import get_greedy, get_beam_search
+from utils.constants import default_ov_config
+from utils.hugging_face import convert_models, get_hugging_face_models, run_hugging_face
 
 TESTS_ROOT = Path(__file__).parent
 
@@ -55,144 +57,6 @@ def get_scheduler_config(scheduler_params: dict = None) -> SchedulerConfig:
     return scheduler_config
 
 
-def convert_to_hf(
-    default_generation_config : HFGenerationConfig,
-    generation_config : GenerationConfig
-) -> HFGenerationConfig:
-    if generation_config is None:
-        return
-
-    kwargs = {}
-    kwargs['return_dict_in_generate'] = True
-
-    # generic parameters
-    kwargs['max_length'] = generation_config.max_length
-    # has higher priority than 'max_length'
-    kwargs['max_new_tokens'] = generation_config.max_new_tokens
-    kwargs['min_new_tokens'] = generation_config.min_new_tokens
-    if generation_config.stop_strings:
-        kwargs['stop_strings'] = generation_config.stop_strings
-
-    # copy default parameters
-    kwargs['bos_token_id'] = default_generation_config.bos_token_id
-    kwargs['pad_token_id'] = default_generation_config.pad_token_id
-
-    if len(generation_config.stop_token_ids) > 0:
-        kwargs['eos_token_id'] = list(generation_config.stop_token_ids)
-    elif generation_config.eos_token_id != -1:
-        kwargs['eos_token_id'] = generation_config.eos_token_id
-    else:
-        kwargs['eos_token_id'] = default_generation_config.eos_token_id
-
-    # copy penalties
-    kwargs['repetition_penalty'] = generation_config.repetition_penalty
-
-    if generation_config.is_beam_search():
-        # beam search case
-        kwargs['num_beam_groups'] = generation_config.num_beam_groups
-        kwargs['num_beams'] = generation_config.num_beams
-        kwargs['length_penalty'] = generation_config.length_penalty
-        kwargs['no_repeat_ngram_size'] = generation_config.no_repeat_ngram_size
-        kwargs['num_return_sequences'] = generation_config.num_return_sequences
-        kwargs['output_scores'] = True
-
-        if generation_config.num_beam_groups > 1:
-            kwargs['diversity_penalty'] = generation_config.diversity_penalty
-
-        # in OpenVINO GenAI this parameter is called stop_criteria,
-        # while in HF it's called early_stopping.
-        # HF values True, False and "never" correspond to OV GenAI values "EARLY", "HEURISTIC" and "NEVER"
-        STOP_CRITERIA_MAP = {
-            StopCriteria.NEVER: "never",
-            StopCriteria.EARLY: True,
-            StopCriteria.HEURISTIC: False
-        }
-
-        kwargs['early_stopping'] = STOP_CRITERIA_MAP[generation_config.stop_criteria]
-    elif generation_config.is_multinomial():
-        # mulitinomial
-        kwargs['temperature'] = generation_config.temperature
-        kwargs['top_k'] = generation_config.top_k
-        kwargs['top_p'] = generation_config.top_p
-        kwargs['do_sample'] = generation_config.do_sample
-    else:
-        # greedy
-        pass
-
-    hf_generation_config = HFGenerationConfig(**kwargs)
-    return hf_generation_config
-
-
-def run_hugging_face(
-    opt_model,
-    hf_tokenizer,
-    prompts: List[str],
-    generation_configs: List[GenerationConfig] | GenerationConfig,
-) -> List[GenerationResult]:
-    generation_results = []
-
-    if type(generation_configs) is list:
-        # process prompt by promp as we have multiple generation configs
-        for prompt, generation_config in zip(prompts, generation_configs):
-            hf_generation_config = convert_to_hf(opt_model.generation_config, generation_config)
-            inputs = {}
-            if hf_tokenizer.chat_template and generation_config.apply_chat_template:
-                prompt = hf_tokenizer.apply_chat_template([{'role': 'user', 'content': prompt}], tokenize=False, add_generation_prompt=True)
-                inputs = hf_tokenizer(prompt, return_tensors="pt", add_special_tokens=False)
-            else:
-                inputs = hf_tokenizer(prompt, return_tensors="pt")
-            input_ids, attention_mask = inputs['input_ids'], inputs['attention_mask']
-            prompt_len = 0 if generation_config.echo else input_ids.numel()
-
-            generate_outputs = opt_model.generate(input_ids=input_ids, attention_mask=attention_mask, generation_config=hf_generation_config, tokenizer=hf_tokenizer)
-            all_text_batch = hf_tokenizer.batch_decode([generated_ids[prompt_len:] for generated_ids in generate_outputs.sequences], skip_special_tokens=True)
-
-            generation_result = GenerationResult()
-            generation_result.m_generation_ids = all_text_batch
-            # sequences_scores are available only for beam search case
-            if generation_config.is_beam_search():
-                generation_result.m_scores = [score for score in generate_outputs.sequences_scores]
-            generation_results.append(generation_result)
-    else:
-        inputs = {}
-        if hf_tokenizer.chat_template and generation_configs.apply_chat_template:
-            processed_prompts = []
-            for prompt in prompts:
-                processed_prompts.append(hf_tokenizer.apply_chat_template([{'role': 'user', 'content': prompt}], tokenize=False, add_generation_prompt=True))
-            # process all prompts as a single batch as we have a single generation config for all prompts
-            inputs = hf_tokenizer(processed_prompts, return_tensors='pt', padding=True, truncation=True, add_special_tokens=False, padding_side='left')
-        else:
-            inputs = hf_tokenizer(prompts, return_tensors='pt', padding=True, truncation=True, padding_side='left')
-        input_ids, attention_mask = inputs['input_ids'], inputs['attention_mask']
-        hf_generation_config = convert_to_hf(opt_model.generation_config, generation_configs)
-        hf_encoded_outputs = opt_model.generate(input_ids, attention_mask=attention_mask, generation_config=hf_generation_config, tokenizer=hf_tokenizer)
-
-        generation_ids = []
-        scores = []
-
-        for idx, hf_encoded_out in enumerate(hf_encoded_outputs.sequences):
-            prompt_idx = idx // hf_generation_config.num_return_sequences
-            prompt_len = 0 if generation_configs.echo else input_ids[prompt_idx].numel()
-            decoded_text = hf_tokenizer.decode(hf_encoded_out[prompt_len:], skip_special_tokens=True)
-            generation_ids.append(decoded_text)
-            if generation_configs.is_beam_search():
-                scores.append(hf_encoded_outputs.sequences_scores[idx])
-
-            # if we need to move to next generation result
-            if (idx + 1) // hf_generation_config.num_return_sequences != prompt_idx:
-                generation_result = GenerationResult()
-                generation_result.m_generation_ids = generation_ids
-                generation_result.m_scores = scores
-                generation_results.append(generation_result)
-                generation_ids = []
-                scores = []
-
-    del hf_tokenizer
-    del opt_model
-
-    return generation_results
-
-
 def run_continuous_batching(
     models_path : Path,
     scheduler_config : SchedulerConfig,
@@ -202,7 +66,7 @@ def run_continuous_batching(
     if type(generation_configs) is not list:
         generation_configs = [generation_configs] * len(prompts)
  
-    cb_pipe = ContinuousBatchingPipeline(models_path, scheduler_config=scheduler_config, device='CPU', tokenizer_properties={}, properties=get_default_properties())
+    cb_pipe = ContinuousBatchingPipeline(models_path, scheduler_config=scheduler_config, device='CPU', tokenizer_properties={}, properties=default_ov_config)
     output = cb_pipe.generate(prompts, generation_configs)
 
     del cb_pipe
@@ -210,15 +74,6 @@ def run_continuous_batching(
 
     return output
 
-
-def get_default_properties():
-    import openvino.properties.hint as hints
-    import openvino as ov
-
-    return {
-        hints.inference_precision : ov.Type.f32,
-        hints.kv_cache_precision : ov.Type.f16,
-    }
 
 def get_models_list_from_path(file_name: str):
     models = []
@@ -259,7 +114,7 @@ def run_llm_pipeline(
     use_cb : bool = False,
     streamer: StreamerWithResults | Callable | StreamerBase = None
 ) -> List[GenerationResult]:
-    properties = get_default_properties()
+    properties = default_ov_config
     if use_cb:
         properties['scheduler_config'] = SchedulerConfig()
     ov_pipe = LLMPipeline(models_path, device='CPU', **properties)
@@ -327,35 +182,6 @@ def compare_generation_results(prompts: List[str], hf_results: List[GenerationRe
     for prompt, ref_result, ov_result, generation_config in zip(prompts, hf_results, ov_results, generation_configs):
         print(f"Prompt = {prompt}\nReference result = {ref_result}\nOpenVINO result = {ov_result.m_generation_ids}")
         compare_generation_result(ref_result, ov_result, generation_config)
-
-
-def get_hugging_face_models(model_id: str):
-    hf_tokenizer = AutoTokenizer.from_pretrained(model_id, trust_remote_code=True)
-    opt_model = OVModelForCausalLM.from_pretrained(model_id, export=True, compile=False, load_in_8bit=False, trust_remote_code=True, ov_config=get_default_properties())
-    return opt_model, hf_tokenizer
-
-
-def convert_models(opt_model : OVModelForCausalLM, hf_tokenizer : AutoTokenizer, models_path: Path):
-    opt_model.save_pretrained(models_path)
-
-    # to store tokenizer config jsons with special tokens
-    hf_tokenizer.save_pretrained(models_path)
-
-    # save generation config
-    opt_model.generation_config.save_pretrained(models_path)
-
-    # convert tokenizers as well
-    convert_and_save_tokenizer(hf_tokenizer, models_path)
-
-
-def convert_and_save_tokenizer(hf_tokenizer : AutoTokenizer, models_path: Path):
-    from openvino_tokenizers import convert_tokenizer
-    from openvino import save_model
-
-    tokenizer, detokenizer = convert_tokenizer(hf_tokenizer, with_detokenizer=True)
-    save_model(tokenizer, models_path / "openvino_tokenizer.xml")
-    save_model(detokenizer, models_path / "openvino_detokenizer.xml")
-
 
 def run_llm_pipeline_with_ref(model_id: str, 
                               prompts: List[str],
