+++ tests/python_tests/ov_genai_test_utils.py
@@ -14,7 +14,9 @@ import shutil
 import json
 
 import openvino_genai as ov_genai
-from common import get_default_properties, delete_rt_info
+from common import delete_rt_info
+
+from utils.constants import default_ov_config
 
 def get_models_list():
     precommit_models = [
@@ -92,7 +94,7 @@ def read_model(params, **tokenizer_kwargs):
 
     if (models_path / "openvino_model.xml").exists():
         opt_model = OVModelForCausalLM.from_pretrained(models_path, trust_remote_code=True,
-                                                       compile=False, device='CPU', ov_config=get_default_properties())
+                                                       compile=False, device='CPU', ov_config=default_ov_config)
     else:
         ov_tokenizer, ov_detokenizer = openvino_tokenizers.convert_tokenizer(hf_tokenizer,
                                                                              with_detokenizer=True,
@@ -104,7 +106,7 @@ def read_model(params, **tokenizer_kwargs):
         hf_tokenizer.save_pretrained(models_path)
 
         opt_model = OVModelForCausalLM.from_pretrained(model_id, export=True, trust_remote_code=True,
-                                                       compile=False, device='CPU', load_in_8bit=False, ov_config=get_default_properties())
+                                                       compile=False, device='CPU', load_in_8bit=False, ov_config=default_ov_config)
         opt_model.generation_config.save_pretrained(models_path)
         opt_model.config.save_pretrained(models_path)
         opt_model.save_pretrained(models_path)
@@ -114,7 +116,7 @@ def read_model(params, **tokenizer_kwargs):
         models_path,
         hf_tokenizer,
         opt_model,
-        ov_genai.LLMPipeline(models_path, 'CPU', ENABLE_MMAP=False, **get_default_properties()),
+        ov_genai.LLMPipeline(models_path, 'CPU', ENABLE_MMAP=False, **default_ov_config),
     )
 
 
@@ -179,7 +181,7 @@ def load_genai_pipe_with_configs(configs: List[Tuple], temp_path):
         with (temp_path / config_name).open('w') as f:
             json.dump(config_json, f)
 
-    ov_pipe = ov_genai.LLMPipeline(temp_path, 'CPU', **get_default_properties())
+    ov_pipe = ov_genai.LLMPipeline(temp_path, 'CPU', **default_ov_config)
 
     for _, config_name in configs:
         os.remove(temp_path / config_name)
@@ -189,4 +191,4 @@ def load_genai_pipe_with_configs(configs: List[Tuple], temp_path):
 
 @functools.lru_cache(1)
 def get_continuous_batching(path):
-    return ov_genai.LLMPipeline(path, 'CPU', scheduler_config=ov_genai.SchedulerConfig(), **get_default_properties())
+    return ov_genai.LLMPipeline(path, 'CPU', scheduler_config=ov_genai.SchedulerConfig(), **default_ov_config)
